Apply text and icon in overlay view
