+++ quickie/src/main/kotlin/io/github/g00fy2/quickie/QROverlayView.kt
@@ -1,6 +1,7 @@
 package io.github.g00fy2.quickie
 
 import android.content.Context
+import android.content.res.Resources.NotFoundException
 import android.graphics.Bitmap
 import android.graphics.Bitmap.Config.ARGB_8888
 import android.graphics.Canvas
@@ -9,6 +10,7 @@ import android.graphics.Paint
 import android.graphics.PorterDuff.Mode.CLEAR
 import android.graphics.PorterDuffXfermode
 import android.graphics.RectF
+import android.graphics.drawable.Drawable
 import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.LayoutInflater
@@ -16,6 +18,7 @@ import android.view.View
 import android.widget.FrameLayout
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.content.ContextCompat
+import androidx.core.content.res.ResourcesCompat
 import androidx.core.graphics.ColorUtils
 import io.github.g00fy2.quickie.databinding.QuickieTextviewBinding
 import kotlin.math.min
@@ -39,11 +42,11 @@ internal class QROverlayView @JvmOverloads constructor(
   private val radius = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, OUT_RADIUS, resources.displayMetrics)
   private val innerRadius =
     TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, OUT_RADIUS - STROKE_WIDTH, resources.displayMetrics)
+  private val titleTextView: AppCompatTextView
   private var maskBitmap: Bitmap? = null
   private var maskCanvas: Canvas? = null
   private var outerFrame = RectF()
   private var innerFrame = RectF()
-  val titleTextView: AppCompatTextView
   var isHighlighted = false
     set(value) {
       field = value
@@ -73,6 +76,28 @@ internal class QROverlayView @JvmOverloads constructor(
     maskCanvas!!.drawRoundRect(innerFrame, innerRadius, innerRadius, transparentPaint)
     canvas.drawBitmap(maskBitmap!!, 0f, 0f, alphaPaint)
     super.onDraw(canvas)
+  }
+
+  fun setCustomTextAndIcon(stringRes: Int, drawableRes: Int) {
+    if (stringRes != 0) {
+      try {
+        titleTextView.setText(stringRes)
+      } catch (ignore: NotFoundException) {
+        // string resource not found
+      }
+    }
+    if (drawableRes != 0) {
+      try {
+        titleTextView.setCompoundDrawables(
+          null,
+          ResourcesCompat.getDrawable(resources, drawableRes, null)?.limitDrawableSize(ICON_DP_MAX_HEIGHT),
+          null,
+          null,
+        )
+      } catch (ignore: NotFoundException) {
+        // drawable resource not found
+      }
+    }
   }
 
   private fun calculateFrameAndTitlePos() {
@@ -113,10 +138,18 @@ internal class QROverlayView @JvmOverloads constructor(
     layoutParams = params
   }
 
+  private fun Drawable.limitDrawableSize(maxDpHeight: Int): Drawable {
+    val scale = (maxDpHeight * resources.displayMetrics.density) / minimumHeight
+    if (scale < 1) setBounds(0, 0, (minimumWidth * scale).roundToInt(), (minimumHeight * scale).roundToInt())
+    else setBounds(0, 0, minimumWidth, minimumHeight)
+    return this
+  }
+
   companion object {
     private const val BACKGROUND_ALPHA = 0.77 * 255
     private const val STROKE_WIDTH = 4f
     private const val OUT_RADIUS = 16f
     private const val FRAME_MARGIN_RATIO = 1f / 4
+    private const val ICON_DP_MAX_HEIGHT = 56
   }
 }
